AttackSystem

+++ Assets/_Game/Scripts/Enemies/Pig.cs
@@ -1,7 +1,7 @@
 using UnityEngine;
 public class Pig : Enemy
 {
-    protected override void TakeDamage(float amount)
+    protected override void TakeDamage(int amount)
     {
         base.TakeDamage(amount);
     }
